RIP custom_data and ZFS :(
For whatever reason, this setup does not work with the new m3.large.x86
machines. For some reason, some of them don't boot into EFI mode,
causing grub to install itself and wreck the partition table. Maybe once
we find a way to require EFI on instances, we'll be able to do bring
ZFS back.